+++ src/commands/Utility/choose.js
@@ -18,8 +18,8 @@ class ChooseCommand extends Command {
         const input = await args.rest('string').catch(() => null);
         if (!input) return commonMessages.sendUsageEmbed(this, message, args);
         const items = input.split(', ');
-        if (items.length < 2) return message.channel.send(`I choose the only possible option: ${items[0]}`);
-        return message.channel.send(`I choose: ${items[Math.floor(Math.random() * items.length)]}`);
+        if (items.length < 2) return message.channel.send(`I choose the only possible option: **${items[0]}**`);
+        return message.channel.send(`I choose: **${items[Math.floor(Math.random() * items.length)]}**`);
     }
 }