+++ src/listeners/ready.js
@@ -11,8 +11,9 @@ class ReadyListener extends Listener {
 
     run(client) {
         client.MutedMembers.sync();
+        client.TimeZones.sync();
         const { username, id } = client.user;
-        this.container.logger.info(`Successfully logged in as ${username} (${id})`);
+        this.container.logger.info(`Successfully logged in as ${username} (${id})!`);
     }
 }